chore: 重构 ChatInput.tsx， 防止重复渲染
