+++ src/panels/ChatPanel/ChatBot/ChatInput/ActionBar/index.tsx
@@ -1,0 +1,15 @@
+import History from './History';
+import Record from './Record';
+import Token from './Token';
+import Voice from './Voice';
+
+const ActionBar = () => (
+  <>
+    <History />
+    <Record />
+    <Voice />
+    <Token />
+  </>
+);
+
+export default ActionBar;
